Conference Centers: Preserve query string parameters across list and edit pages (#7886)
* Conference Centers: Preserve query string parameters across list and edit pages

* Update conference_center_edit.php

* Update conference_rooms.php

* Update conference_room_edit.php

+++ app/conference_centers/conference_centers.php
@@ -17,7 +17,7 @@
 
 	The Initial Developer of the Original Code is
 	Mark J Crane <markjcrane@fusionpbx.com>
-	Portions created by the Initial Developer are Copyright (C) 2008-2024
+	Portions created by the Initial Developer are Copyright (C) 2008-2026
 	the Initial Developer. All Rights Reserved.
 
 	Contributor(s):
@@ -39,8 +39,32 @@
 	$language = new text;
 	$text = $language->get();
 
-//set additional variables
-	$show = $_GET["show"] ?? '';
+// Set variables from GET parameters
+	$page = is_numeric($_GET['page'] ?? '') ? $_GET['page'] : 0;
+	$order_by = preg_replace('#[^a-zA-Z0-9_\-]#', '', ($_GET['order_by'] ?? ''));
+	$order = ($_GET['order'] ?? '') === 'desc' ? 'desc' : 'asc';
+	$sort = $order_by == 'conference_center_extension' ? 'natural' : null;
+	$search = $_GET['search'] ?? '';
+	$show = $_GET['show'] ?? '';
+
+// Build the query string
+	$param = [];
+	if (!empty($page)) {
+		$param['page'] = $page;
+	}
+	if (!empty($_GET['order_by'])) {
+		$param['order_by'] = $order_by;
+	}
+	if (!empty($_GET['order'])) {
+		$param['order'] = $order;
+	}
+	if (!empty($search)) {
+		$param['search'] = $search;
+	}
+	if (!empty($show) && $show == 'all' && permission_exists('conference_center_all')) {
+		$param['show'] = $show;
+	}
+	$query_string = http_build_query($param);
 
 //set from session variables
 	$list_row_edit_button = $settings->get('theme', 'list_row_edit_button', false);
@@ -48,7 +72,6 @@
 //get posted data
 	if (!empty($_POST['conference_centers'])) {
 		$action = $_POST['action'];
-		$search = $_POST['search'] ?? '';
 		$conference_centers = $_POST['conference_centers'];
 	}
 
@@ -77,17 +100,11 @@
 				break;
 		}
 
-		header('Location: conference_centers.php'.(!empty($search) ? '?search='.urlencode($search) : ''));
+		header('Location: conference_centers.php'.($query_string ? '?'.$query_string : ''));
 		exit;
 	}
 
-//get variables used to control the order
-	$order_by = $_GET["order_by"] ?? '';
-	$order = $_GET["order"] ?? '';
-	$sort = $order_by == 'conference_center_extension' ? 'natural' : null;
-
 //add the search term
-	$search = strtolower($_GET["search"] ?? '');
 	if (!empty($search)) {
 		$sql_search = "and ( ";
 		$sql_search .= "lower(conference_center_name) like :search ";
@@ -95,7 +112,7 @@
 		$sql_search .= "or lower(conference_center_greeting) like :search ";
 		$sql_search .= "or lower(conference_center_description) like :search ";
 		$sql_search .= ") ";
-		$parameters['search'] = '%'.$search.'%';
+		$parameters['search'] = '%'.lower_case($search).'%';
 	}
 
 //prepare to page the results
@@ -110,14 +127,8 @@
 
 //prepare to page the results
 	$rows_per_page = $settings->get('domain', 'paging', 50);
-	$param = "&search=".urlencode($search);
-	if ($show == "all" && permission_exists('conference_center_all')) {
-		$param .= "&show=all";
-	}
-	$page = $_GET['page'] ?? '';
-	if (empty($page)) { $page = 0; $_GET['page'] = 0; }
-	list($paging_controls, $rows_per_page) = paging($num_rows, $param, $rows_per_page);
-	list($paging_controls_mini, $rows_per_page) = paging($num_rows, $param, $rows_per_page, true);
+	list($paging_controls, $rows_per_page) = paging($num_rows, $query_string, $rows_per_page);
+	list($paging_controls_mini, $rows_per_page) = paging($num_rows, $query_string, $rows_per_page, true);
 	$offset = $rows_per_page * $page;
 
 //get the list
@@ -166,7 +177,7 @@
 	}
 	echo button::create(['type'=>'button','label'=>$text['button-rooms'],'icon'=>'door-open','style'=>'margin-right: 15px;','link'=>'conference_rooms.php']);
 	if (permission_exists('conference_center_add')) {
-		echo button::create(['type'=>'button','label'=>$text['button-add'],'icon'=>$settings->get('theme', 'button_icon_add'),'id'=>'btn_add','link'=>'conference_center_edit.php']);
+		echo button::create(['type'=>'button','label'=>$text['button-add'],'icon'=>$settings->get('theme', 'button_icon_add'),'id'=>'btn_add','link'=>'conference_center_edit.php'.($query_string ? '?'.$query_string : '')]);
 	}
 	if (permission_exists('conference_center_edit') && $conference_centers) {
 		echo button::create(['type'=>'button','label'=>$text['button-toggle'],'icon'=>$settings->get('theme', 'button_icon_toggle'),'id'=>'btn_toggle','name'=>'btn_toggle','style'=>'display: none;','onclick'=>"modal_open('modal-toggle','btn_toggle');"]);
@@ -175,13 +186,13 @@
 		echo button::create(['type'=>'button','label'=>$text['button-delete'],'icon'=>$settings->get('theme', 'button_icon_delete'),'id'=>'btn_delete','name'=>'btn_delete','style'=>'display: none;','onclick'=>"modal_open('modal-delete','btn_delete');"]);
 	}
 	echo 		"<form id='form_search' class='inline' method='get'>\n";
-	if (permission_exists('conference_center_all')) {
-		if ($show == 'all') {
-			echo "		<input type='hidden' name='show' value='all'>";
+	foreach ($param as $key => $value) {
+		if ($key !== 'search' && $key !== 'page') {
+			echo "		<input type='hidden' name='".escape($key)."' value='".escape($value)."'>\n";
 		}
-		else {
-			echo button::create(['type'=>'button','label'=>$text['button-show_all'],'icon'=>$settings->get('theme', 'button_icon_all'),'link'=>'?type=&show=all'.(!empty($search) ? "&search=".urlencode($search) : null)]);
-		}
+	}
+	if ($show !== 'all' && permission_exists('conference_center_all')) {
+		echo button::create(['type'=>'button','label'=>$text['button-show_all'],'icon'=>$settings->get('theme', 'button_icon_all'),'link'=>'?type=&show=all'.(!empty($search) ? "&search=".urlencode($search) : null)]);
 	}
 	echo 		"<input type='text' class='txt list-search' name='search' id='search' value=\"".escape($search)."\" placeholder=\"".$text['label-search']."\" onkeydown=''>";
 	echo button::create(['label'=>$text['button-search'],'icon'=>$settings->get('theme', 'button_icon_search'),'type'=>'submit','id'=>'btn_search']);
@@ -206,7 +217,6 @@
 
 	echo "<form id='form_list' method='post'>\n";
 	echo "<input type='hidden' id='action' name='action' value=''>\n";
-	echo "<input type='hidden' name='search' value=\"".escape($search)."\">\n";
 
 	echo "<div class='card'>\n";
 	echo "<table class='list'>\n";
@@ -217,14 +227,14 @@
 		echo "	</th>\n";
 	}
 	if ($show == "all" && permission_exists('conference_center_all')) {
-		echo th_order_by('domain_name', $text['label-domain'], $order_by, $order, $param, "class='shrink'");
+		echo th_order_by('domain_name', $text['label-domain'], $order_by, $order, null, "class='shrink'", $query_string);
 	}
-	echo th_order_by('conference_center_name', $text['label-conference_center_name'], $order_by, $order);
-	echo th_order_by('conference_center_extension', $text['label-conference_center_extension'], $order_by, $order);
-	echo th_order_by('conference_center_greeting', $text['label-conference_center_greeting'], $order_by, $order);
-	echo th_order_by('conference_center_pin_length', $text['label-conference_center_pin_length'], $order_by, $order, null, "class='center shrink'");
-	echo th_order_by('conference_center_enabled', $text['label-conference_center_enabled'], $order_by, $order, null, "class='center'");
-	echo th_order_by('conference_center_description', $text['label-conference_center_description'], $order_by, $order, null, "class='hide-sm-dn'");
+	echo th_order_by('conference_center_name', $text['label-conference_center_name'], $order_by, $order, null, null, $query_string);
+	echo th_order_by('conference_center_extension', $text['label-conference_center_extension'], $order_by, $order, null, null, $query_string);
+	echo th_order_by('conference_center_greeting', $text['label-conference_center_greeting'], $order_by, $order, null, null, $query_string);
+	echo th_order_by('conference_center_pin_length', $text['label-conference_center_pin_length'], $order_by, $order, null, "class='center shrink'", $query_string);
+	echo th_order_by('conference_center_enabled', $text['label-conference_center_enabled'], $order_by, $order, null, "class='center'", $query_string);
+	echo th_order_by('conference_center_description', $text['label-conference_center_description'], $order_by, $order, null, "class='hide-sm-dn'", $query_string);
 	if (permission_exists('conference_center_edit') && $list_row_edit_button) {
 		echo "	<td class='action-button'>&nbsp;</td>\n";
 	}
@@ -235,7 +245,7 @@
 		foreach ($conference_centers as $row) {
 			$list_row_url = '';
 			if (permission_exists('conference_center_edit')) {
-				$list_row_url = "conference_center_edit.php?id=".$row['conference_center_uuid'];
+				$list_row_url = "conference_center_edit.php?id=".$row['conference_center_uuid'].($query_string ? '&'.$query_string : '');
 				if ($row['domain_uuid'] != $_SESSION['domain_uuid'] && permission_exists('domain_select')) {
 					$list_row_url .= '&domain_uuid='.urlencode($row['domain_uuid']).'&domain_change=true';
 				}
